requirements.txt
numpy==1.23.0
chumpy==0.70
scikit-image==0.18.1
lpips==0.1.4
pytorch-lightning<2
torchgeometry==0.1.2
omegaconf==2.2.3
nerfacc==0.5.3
scipy==1.11.3
wandb==0.15.4
hydra-core==1.3.2
h5py==3.10.0
trimesh
matplotlib
opencv-python
imageio
imageio-ffmpeg
PyMCubes
pyransac3d
torch_efficient_distloss
tensorboard
ninja
pyexr
pytest
